feat: add tests and fix readme

+++ goals/serializers.py
@@ -123,26 +123,28 @@ class BoardSerializer(serializers.ModelSerializer):
         read_only_fields = ("id", "created", "updated")
 
     def update(self, instance, validated_data):
-        owner = validated_data.pop("user")
-        new_participants = validated_data.pop("participants")
-        new_by_id = {part["user"].id: part for part in new_participants}
+        owner = self.context["request"].user #validated_data.pop("user")
+        if "participants" in validated_data:
+            new_participants = validated_data.get("participants", [])
+            new_by_id = {part["user"].id: part for part in new_participants}
 
-        old_participants = instance.board_participants.exclude(user=owner)
-        with transaction.atomic():
-            for old_participant in old_participants:
-                if old_participant.user_id not in new_by_id:
-                    old_participant.delete()
-                else:
-                    if old_participant.role != new_by_id[old_participant.user_id]["role"]:
-                        old_participant.role = new_by_id[old_participant.user_id]["role"]
-                        old_participant.save()
-                    new_by_id.pop(old_participant.user_id)
-                for new_part in new_by_id.values():
-                    BoardParticipant.objects.create(
-                        board=instance, user=new_part["user"], role=new_part["role"]
-                    )
+            old_participants = instance.board_participants.exclude(user=owner)
+            with transaction.atomic():
+                for old_participant in old_participants:
+                    if old_participant.user_id not in new_by_id:
+                        old_participant.delete()
+                    else:
+                        if old_participant.role != new_by_id[old_participant.user_id]["role"]:
+                            old_participant.role = new_by_id[old_participant.user_id]["role"]
+                            old_participant.save()
+                        new_by_id.pop(old_participant.user_id)
+                    for new_part in new_by_id.values():
+                        BoardParticipant.objects.create(
+                            board=instance, user=new_part["user"], role=new_part["role"]
+                        )
 
-        instance.title = validated_data["title"]
+        if "title" in validated_data:
+            instance.title = validated_data["title"]
         instance.save()
 
         return instance
